refactor(effect): split shader D-Bus config, texture cache, and ctor wiring
beginShaderTransition's GL compile stage becomes compileOrLoadAnimationShader;
nested state types move to effect_state.h.
