+++ kwin-effect/plasmazoneseffect/shader_transitions.cpp
@@ -68,12 +68,9 @@ namespace PlasmaZones {
 
 Q_DECLARE_LOGGING_CATEGORY(lcEffect)
 
-using ShaderInternal::injectKwinDefineAfterVersion;
-using ShaderInternal::kCustomColorsElementNames;
-using ShaderInternal::kCustomParamsElementNames;
-using ShaderInternal::kITextureResolutionKeys;
 using ShaderInternal::kUserTextureSamplerNames;
 using ShaderInternal::kUserTextureWrapKeys;
+using ShaderInternal::loadUserTextureImage;
 using ShaderInternal::shaderClockNowMs;
 
 /// Splice `#define PLASMAZONES_KWIN` between the shader's `#version`
@@ -241,50 +238,6 @@ QByteArray ShaderInternal::injectKwinDefineAfterVersion(const QString& source)
 
 namespace {
 
-/// Load a user-texture file into a QImage. Mirrors the daemon-side path
-/// in `PhosphorRendering::ShaderEffect::setShaderParams`: PNG/JPG/etc.
-/// load via `QImage`; `.svg` / `.svgz` rasterise via `QSvgRenderer` at
-/// the requested max-axis dimension (defaulting to 1024 to match
-/// ZoneShaderItem's pre-unification SVG size). The `QImage::Format_RGBA8888`
-/// conversion ensures `KWin::GLTexture::upload` always sees a consistent
-/// pixel layout regardless of the source file's native format. Returns
-/// a null QImage on any failure; the caller logs and skips the slot.
-///
-/// SVG rasterise target: `Format_ARGB32_Premultiplied` first, then
-/// convert to `Format_RGBA8888` for the GPU upload. QPainter's
-/// source-over compositing is defined for premultiplied targets;
-/// rendering an SVG with semi-transparent strokes/fills directly into
-/// `Format_RGBA8888` produces subtly wrong alpha at partial-cover paths.
-/// Matches the daemon path's behaviour exactly so the same SVG renders
-/// identically on both runtimes.
-inline QImage loadUserTextureImage(const QString& path, int svgMaxDim = 1024)
-{
-    if (path.isEmpty()) {
-        return {};
-    }
-    const bool isSvg = path.endsWith(QLatin1String(".svg"), Qt::CaseInsensitive)
-        || path.endsWith(QLatin1String(".svgz"), Qt::CaseInsensitive);
-    if (isSvg) {
-        QSvgRenderer renderer(path);
-        if (!renderer.isValid()) {
-            return {};
-        }
-        QSize size = renderer.defaultSize();
-        if (!size.isEmpty()) {
-            size.scale(svgMaxDim, svgMaxDim, Qt::KeepAspectRatio);
-        } else {
-            size = QSize(svgMaxDim, svgMaxDim);
-        }
-        QImage rasterised(size, QImage::Format_ARGB32_Premultiplied);
-        rasterised.fill(Qt::transparent);
-        QPainter painter(&rasterised);
-        renderer.render(&painter);
-        painter.end();
-        return rasterised.convertToFormat(QImage::Format_RGBA8888);
-    }
-    return QImage(path).convertToFormat(QImage::Format_RGBA8888);
-}
-
 /// Translate a metadata / params wrap-mode string to the GL enum the
 /// kwin-effect applies at bind time. Empty / unrecognised values fall
 /// through to `GL_CLAMP_TO_EDGE` (the GL default and the daemon's
@@ -299,283 +252,7 @@ inline GLenum wrapStringToEnum(const QString& wrap)
     return GL_CLAMP_TO_EDGE;
 }
 
-/// Parse a D-Bus setting variant containing a JSON-encoded string and
-/// dispatch to one of two callers based on the document's top-level
-/// shape. Used by the three `load*FromDbus` setting fetchers in
-/// `shader_transitions.cpp` — `loadShaderProfileFromDbus`,
-/// `loadMotionProfileTreeFromDbus`, `loadShaderRegistryFromDbus`. Each
-/// loader differs only in (a) which shape it expects and (b) what it
-/// does with the parsed JSON, so every other piece (UTF-8 decode,
-/// document-shape check, malformed-payload warning text) collapses
-/// into a single helper call. `loadRuleAnimationsFromDbus` is the
-/// odd one out — it issues a raw `QDBusMessage::createMethodCall` to
-/// `getAllRules` and parses with `QJsonDocument::fromJson` directly,
-/// because it slices the parsed rules through
-/// `excludeRulesFrom` / `excludeAnimationsRulesFrom` before sinking.
-///
-/// The `name` argument feeds the warning so the failure site is
-/// identifiable in journals; pass the same `SettingProperty` constant
-/// the loader requested.
-///
-/// `objectSink` runs when the document is a top-level JSON object;
-/// `arraySink` runs when it is a top-level JSON array. Pass a
-/// no-op (empty std::function) for the shape the caller doesn't
-/// expect — a payload of the wrong shape logs and is dropped.
-inline void dispatchJsonSetting(QLatin1String name, const QVariant& v,
-                                std::function<void(const QJsonObject&)> objectSink,
-                                std::function<void(const QJsonArray&)> arraySink)
-{
-    const QJsonDocument doc = QJsonDocument::fromJson(v.toString().toUtf8());
-    if (doc.isObject() && objectSink) {
-        objectSink(doc.object());
-    } else if (doc.isArray() && arraySink) {
-        arraySink(doc.array());
-    } else {
-        // Name the expected shape explicitly from which sink the caller
-        // wired — covers all four combinations (object-only, array-only,
-        // both, neither). Picking from the truthy ternary would lie when
-        // both sinks are bound, or when neither is.
-        const char* expected = (objectSink && arraySink) ? "object or array"
-            : objectSink                                 ? "object"
-            : arraySink                                  ? "array"
-                                                         : "(no shape — caller wired neither sink)";
-        qCWarning(lcEffect) << "Failed to parse" << name << "from D-Bus — payload is not a JSON" << expected;
-    }
-}
-
 } // namespace
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Async texture pre-warm.
-//
-// Pattern: a `QRunnable` posted to `m_shaderManager.m_textureLoaderPool` performs the
-// CPU-bound load (`loadUserTextureImage` — QImage decode for raster
-// formats, QSvgRenderer rasterise for SVG/SVGZ) on a worker thread,
-// producing a `QImage` in `Format_RGBA8888`. The worker then dispatches
-// a queued slot back to `this` via `QMetaObject::invokeMethod(...,
-// Qt::QueuedConnection)` so the GL upload (`KWin::GLTexture::upload`)
-// runs on the GL-context thread (the compositor thread). The cache
-// insert and in-flight set bookkeeping happen entirely on the
-// compositor thread, so no locking is needed against
-// `m_shaderManager.m_textureCache` or `m_shaderManager.m_textureLoadsInFlight`.
-//
-// Thread-safety notes:
-//   • The worker only reads the captured path string (`m_path`) and
-//     `m_svgMaxDim` — both POD captured-by-value at submission time.
-//     It NEVER touches `m_shaderManager.m_textureCache` or `m_shaderManager.m_textureLoadsInFlight`;
-//     all access to those members happens on the compositor thread,
-//     either at submission time or inside the queued upload lambda.
-//     The submission-time generation captured into the worker lets
-//     the queued lambda detect a hot-reload that cleared the cache
-//     and discard the upload before touching `m_shaderManager.m_textureLoadsInFlight`.
-//   • `QSvgRenderer` is NOT thread-safe across instances (it owns
-//     mutable rasteriser state during `render()` per Qt docs).
-//     `loadUserTextureImage` constructs a fresh `QSvgRenderer` per
-//     call, so each worker invocation gets its own renderer — safe.
-//   • `QImage(path)` (PNG/JPG/etc. decode) is thread-safe across
-//     instances per Qt docs.
-//   • `KWin::GLTexture::upload` MUST run on the GL thread; that's
-//     why the upload is dispatched back via QueuedConnection rather
-//     than completed inline on the worker.
-// ─────────────────────────────────────────────────────────────────────────────
-void PlasmaZonesEffect::evictLruTextureIfOverBound(const ShaderTransition* pending)
-{
-    // Build the set of cache pointers currently referenced by any
-    // active transition's userTextures slots ONCE. Eviction skips
-    // every one of these — the transition holds a raw non-owning
-    // pointer that would dangle if we erased the entry. The set
-    // doesn't change between iterations because the eviction below
-    // only removes NON-in-flight entries; the set of in-flight
-    // pointers is invariant across the loop, so we hoist the build.
-    std::unordered_set<const CachedTexture*> inFlight;
-    for (const auto& [_, transition] : m_shaderManager.shaderTransitions()) {
-        for (CachedTexture* tex : transition.userTextures) {
-            if (tex) {
-                inFlight.insert(tex);
-            }
-        }
-    }
-    // The transition still being BUILT is not in shaderTransitions() yet (it is
-    // inserted once every slot is resolved), so the slots it has already filled would
-    // otherwise be evictable by the sweep its own next slot triggers — leaving it
-    // holding a dangling pointer. It survives today only by arithmetic: the soft bound
-    // is many times the slot count and every entry it touches carries the newest access
-    // tick, so it can never be the victim. That is a property of two constants, not a
-    // guarantee, and neither constant knows it is load-bearing.
-    if (pending) {
-        for (CachedTexture* tex : pending->userTextures) {
-            if (tex) {
-                inFlight.insert(tex);
-            }
-        }
-    }
-    while (m_shaderManager.m_textureCache.size() > ShaderTransitionManager::kTextureCacheSoftBound) {
-        // Find the cache entry with the smallest lastAccessTick that is
-        // NOT in-flight. If every entry is in flight (pathological;
-        // would require >ShaderTransitionManager::kTextureCacheSoftBound concurrent transitions
-        // each referencing a unique texture), break — the cache
-        // transiently exceeds the bound rather than tearing a live
-        // pointer. Self-heals on the next eviction once a transition
-        // ends.
-        auto evictIt = m_shaderManager.m_textureCache.end();
-        quint64 oldestTick = std::numeric_limits<quint64>::max();
-        for (auto it = m_shaderManager.m_textureCache.begin(); it != m_shaderManager.m_textureCache.end(); ++it) {
-            if (inFlight.count(&it->second) > 0) {
-                continue;
-            }
-            if (it->second.lastAccessTick < oldestTick) {
-                oldestTick = it->second.lastAccessTick;
-                evictIt = it;
-            }
-        }
-        if (evictIt == m_shaderManager.m_textureCache.end()) {
-            return; // every entry is in flight; no safe eviction this pass
-        }
-        qCDebug(lcEffect) << "evictLruTextureIfOverBound: evicting" << evictIt->first
-                          << "(lastAccessTick=" << evictIt->second.lastAccessTick
-                          << ", cache size=" << m_shaderManager.m_textureCache.size() << ")";
-        m_shaderManager.m_textureCache.erase(evictIt);
-    }
-}
-
-void PlasmaZonesEffect::warmUserTextureAsync(const QString& absolutePath)
-{
-    if (absolutePath.isEmpty()) {
-        return;
-    }
-    // Already warm — fast path, no allocation.
-    if (m_shaderManager.m_textureCache.find(absolutePath) != m_shaderManager.m_textureCache.end()) {
-        return;
-    }
-    // Already in flight — a worker is mid-load; deduplicate to avoid
-    // duplicate GPU uploads when several transitions request the
-    // same path before the first one completes.
-    if (m_shaderManager.m_textureLoadsInFlight.contains(absolutePath)) {
-        return;
-    }
-    m_shaderManager.m_textureLoadsInFlight.insert(absolutePath);
-
-    // SVG default size matches `loadUserTextureImage`'s 1024 max-axis.
-    // Captured by value into the worker. The cache is path-keyed; if
-    // we ever need per-asset size variants the cache key must include
-    // the rasterised dimension, otherwise two callers requesting the
-    // same SVG at different sizes would race on whichever one wins.
-    constexpr int svgMaxDim = 1024;
-
-    // Capture the cache generation at submission time. The queued
-    // upload lambda compares this against the live
-    // `m_shaderManager.m_textureCacheGeneration` and discards if mismatched — i.e. a
-    // hot-reload (`effectsChanged`) bumped the generation between
-    // submission and upload, so this worker's bytes are stale and
-    // must not re-populate the cleared cache.
-    const quint64 submissionGeneration = m_shaderManager.m_textureCacheGeneration;
-
-    class Loader : public QRunnable
-    {
-    public:
-        Loader(QPointer<PlasmaZonesEffect> effect, QString path, int svgMaxDim, quint64 submissionGeneration)
-            : m_effect(std::move(effect))
-            , m_path(std::move(path))
-            , m_svgMaxDim(svgMaxDim)
-            , m_submissionGeneration(submissionGeneration)
-        {
-        }
-        void run() override
-        {
-            QImage img = loadUserTextureImage(m_path, m_svgMaxDim);
-            QPointer<PlasmaZonesEffect> effect = m_effect;
-            QString path = m_path;
-            const quint64 submissionGeneration = m_submissionGeneration;
-            // Bounce back to the compositor thread for the GL upload.
-            // The QPointer guards against the effect being destroyed
-            // while the worker was running — destructor's
-            // `m_shaderManager.m_textureLoaderPool.waitForDone()` already protects
-            // against this for the in-process teardown case, but the
-            // QPointer is defence-in-depth for any future caller that
-            // schedules this without owning the pool's lifetime.
-            QMetaObject::invokeMethod(
-                effect.data(),
-                [effect, path, img = std::move(img), submissionGeneration]() mutable {
-                    if (!effect) {
-                        return;
-                    }
-                    // Generation check FIRST — before touching
-                    // `m_shaderManager.m_textureCache` or `m_shaderManager.m_textureLoadsInFlight`. If
-                    // the cache was cleared underneath us by a hot-
-                    // reload (`effectsChanged`) the in-flight set was
-                    // already cleared too; touching it now would mean
-                    // racing with state the lambda has no business
-                    // mutating. Discard cleanly.
-                    if (submissionGeneration != effect->m_shaderManager.m_textureCacheGeneration) {
-                        qCDebug(lcEffect) << "warmUserTextureAsync: discarding stale upload for" << path
-                                          << "(generation mismatch — cache cleared during load)";
-                        return;
-                    }
-                    effect->m_shaderManager.m_textureLoadsInFlight.remove(path);
-                    if (img.isNull()) {
-                        qCWarning(lcEffect) << "warmUserTextureAsync: load failed for" << path;
-                        return;
-                    }
-                    // This lands from the EVENT LOOP, between frames — not inside a
-                    // paint cycle — so the compositor's GL context is not guaranteed
-                    // current. Everything below is GL: the upload, the filter/wrap
-                    // state, and the LRU sweep, whose eviction destroys a GLTexture and
-                    // therefore calls glDeleteTextures. Issuing any of that against no
-                    // context (or, worse, another thread's) is undefined. Every other
-                    // off-paint GL site in the effect makes the context current first
-                    // and says so; this one was missed. Bail rather than latch: the next
-                    // use of the path loads it synchronously inside a paint.
-                    if (!KWin::effects || !KWin::effects->makeOpenGLContextCurrent()) {
-                        qCWarning(lcEffect)
-                            << "warmUserTextureAsync: no current GL context, discarding upload for" << path;
-                        return;
-                    }
-                    // Re-check the cache: another transition may have
-                    // synchronously loaded this path while we were on
-                    // the worker. Honour the existing entry; dropping
-                    // ours avoids a redundant GPU upload.
-                    if (effect->m_shaderManager.m_textureCache.find(path)
-                        != effect->m_shaderManager.m_textureCache.end()) {
-                        return;
-                    }
-                    std::unique_ptr<KWin::GLTexture> gpuTex = KWin::GLTexture::upload(img);
-                    if (!gpuTex) {
-                        qCWarning(lcEffect) << "warmUserTextureAsync: GLTexture::upload failed for" << path;
-                        return;
-                    }
-                    gpuTex->setFilter(GL_LINEAR);
-                    gpuTex->setWrapMode(GL_CLAMP_TO_EDGE);
-                    CachedTexture cachedTex;
-                    cachedTex.texture = std::move(gpuTex);
-                    cachedTex.lastAppliedWrap = GL_CLAMP_TO_EDGE;
-                    cachedTex.lastAccessTick = ++effect->m_shaderManager.m_textureCacheAccessTick;
-                    effect->m_shaderManager.m_textureCache.emplace(path, std::move(cachedTex));
-                    // No `pending` to guard: this is a WARM, not a bind. Nothing holds a
-                    // pointer to the entry we just inserted, so the worst the sweep can do
-                    // is discard the warm we just paid for — which is exactly what being
-                    // over the bound means. A live transition's slots are protected by the
-                    // shaderTransitions() scan inside.
-                    effect->evictLruTextureIfOverBound(/*pending=*/nullptr);
-                    qCDebug(lcEffect) << "warmUserTextureAsync: cached" << path;
-                },
-                Qt::QueuedConnection);
-        }
-
-    private:
-        QPointer<PlasmaZonesEffect> m_effect;
-        QString m_path;
-        int m_svgMaxDim;
-        quint64 m_submissionGeneration;
-    };
-
-    // Pass `this` as a QPointer so the conversion happens at the call
-    // site (where `this` is known live), not inside the ctor body where
-    // a freed `effect` mid-construction would silently degrade to a raw
-    // pointer that never registers with QPointer's tracker.
-    auto* loader = new Loader(QPointer<PlasmaZonesEffect>(this), absolutePath, svgMaxDim, submissionGeneration);
-    loader->setAutoDelete(true);
-    m_shaderManager.m_textureLoaderPool.start(loader);
-}
 
 bool PlasmaZonesEffect::beginShaderTransition(KWin::EffectWindow* window,
                                               const PhosphorAnimationShaders::ShaderProfile& profile, int durationMs,
@@ -691,388 +368,23 @@ bool PlasmaZonesEffect::beginShaderTransition(KWin::EffectWindow* window,
     // update that turns around at one of the guards above without touching GL at all.
     ensureGlContextCurrent();
 
-    // KWin-specific default vertex stage. Hardcoded here rather than
-    // loaded from `data/animations/shared/animation.vert` because that
-    // file is shared with the daemon's RHI surface pipeline, which
-    // requires all uniforms to live in UBOs (default-block uniforms
-    // aren't supported under Qt-RHI/SPIR-V) and supplies vertex
-    // positions already in clip space — exactly the opposite of what
-    // KWin's classic-GL OffscreenEffect needs:
-    //
-    //   • Positions arrive in screen-pixel space (KWin's
-    //     `OffscreenData::paint` writes window-rect pixel coords into
-    //     the streaming buffer at `GLVertex2DLayout`'s position slot).
-    //   • The pixel→NDC projection lives in the
-    //     `modelViewProjectionMatrix` default-block uniform, which KWin
-    //     sets via `Mat4Uniform::ModelViewProjectionMatrix` (mapped to
-    //     uniform name `modelViewProjectionMatrix` per
-    //     `KWin::GLShader::resolveLocations` in
-    //     `<opengl/glshader.cpp>`). Skipping the multiplication leaves
-    //     the redirected quad at coords like (1920, 1080) — entirely
-    //     outside the [-1, 1] viewport — and the transition shader
-    //     runs but paints to nothing. Manifest: `tryBeginShader`
-    //     resolves, `beginShaderTransition` installs cleanly, no
-    //     compile warnings, but the user sees no visible animation.
-    //   • Attribute slot indices match `KWin::VA_Position` (0) and
-    //     `KWin::VA_TexCoord` (1) per `<opengl/glvertexbuffer.h>`;
-    //     explicit `layout(location = N)` decorations bypass KWin's
-    //     `bindAttributeLocation("position", ...)` lookup (which is
-    //     name-only and would mismatch our `texCoord` vs KWin's
-    //     `texcoord`).
-    //
-    // texCoord is flipped to a Y-down screen UV. KWin's
-    // `OffscreenData::paint` populates the texCoord attribute from its
-    // bottom-origin offscreen FBO (Y-up). The shader contract — and the
-    // daemon's `kDefaultVertexShaderSource` — expect `vTexCoord` Y-down
-    // (y = 0 at the top), so flip here. The redirected window texture
-    // is Y-up to match KWin's FBO, which is why shaders read it through
-    // `surfaceColor()` (which re-flips the sample coordinate) instead
-    // of sampling `uTexture0` directly — see the canonical header
-    // `data/animations/shared/animation_uniforms.glsl`. Without the
-    // flip, `vTexCoord` is Y-up on this path: the window still composes
-    // upright (the Y-up texture cancels it) but every screen-space
-    // effect — matrix rain direction, directional wipes — renders
-    // inverted. The other kwin-vs-daemon difference is `gl_Position`:
-    // KWin needs the modelViewProjectionMatrix to place the redirected
-    // quad, the daemon emits clip-space directly.
-    //
-    // Authors that ship a per-shader vertex stage via metadata's
-    // `vertexShader` field own the matrix / attribute contract
-    // themselves: emit the same `vTexCoord` flip under
-    // `#ifdef PLASMAZONES_KWIN` and multiply `position` by
-    // modelViewProjectionMatrix on the kwin path. See the canonical
-    // GLSL header (`data/animations/shared/animation_uniforms.glsl`).
-    static const QByteArray kKwinDefaultVertexSource = QByteArrayLiteral(
-        "#version 450\n"
-        "\n"
-        "layout(location = 0) in vec2 position;\n"
-        "layout(location = 1) in vec2 texCoord;\n"
-        "\n"
-        "layout(location = 0) out vec2 vTexCoord;\n"
-        "\n"
-        "uniform mat4 modelViewProjectionMatrix;\n"
-        "\n"
-        "void main() {\n"
-        "    vTexCoord = vec2(texCoord.x, 1.0 - texCoord.y);\n"
-        "    gl_Position = modelViewProjectionMatrix * vec4(position, 0.0, 1.0);\n"
-        "}\n");
-
-    auto cacheIt = m_shaderManager.m_shaderCache.find(effectId);
-    if (cacheIt == m_shaderManager.m_shaderCache.end()) {
-        // Diagnostic-once-per-compile: log multipass degradation when the
-        // shader is first compiled for this session, not on every transition
-        // install. Lifecycle events (window.move on a drag, window.focus on
-        // alt-tab) can fire beginShaderTransition many times in quick
-        // succession against an already-cached effect; a per-install log
-        // would flood the journal. Cache invalidation (effectsChanged →
-        // m_shaderManager.m_shaderCache.clear) re-fires the log at the next install, which
-        // is the right semantic for hot-reload.
-        if (eff.isMultipass) {
-            qCInfo(lcEffect) << "Animation effect" << effectId
-                             << "is multipass — compositor path runs single-pass only (buffer passes skipped)";
-        }
-
-        QFile shaderFile(eff.fragmentShaderPath);
-        if (!shaderFile.open(QIODevice::ReadOnly)) {
-            qCWarning(lcEffect) << "Failed to open shader file" << eff.fragmentShaderPath;
-            return false;
-        }
-        const QString rawSource = QString::fromUtf8(shaderFile.readAll());
-        if (rawSource.isEmpty()) {
-            qCWarning(lcEffect) << "Shader file is empty" << eff.fragmentShaderPath;
-            return false;
-        }
-        QStringList animIncludePaths;
-        for (const QString& sp : m_shaderManager.m_animationShaderRegistry.searchPaths()) {
-            const QString sharedDir = sp + QStringLiteral("/shared");
-            if (QDir(sharedDir).exists()) {
-                animIncludePaths.append(sharedDir);
-            }
-        }
-        QString includeError;
-        const QString currentDir = QFileInfo(eff.fragmentShaderPath).absolutePath();
-        // T1.5: assemble an entry-only animation pack (pTransition / pIn+pOut,
-        // no main()) into a full translation unit BEFORE expansion — identical to
-        // the daemon's loadFragmentShader — so the prologue's `#include` resolves
-        // and the generated main() dispatches by direction. A traditional main()
-        // pack is returned unchanged. Same prologue + candidates the daemon uses,
-        // so both runtimes compile the same source.
-        const QString assembledSource = PhosphorShaders::assembleEntryPoint(
-            rawSource, PhosphorAnimationShaders::AnimationShaderRegistry::animationEntryPrologue(),
-            PhosphorAnimationShaders::AnimationShaderRegistry::animationEntryCandidates());
-        QString expanded = PhosphorShaders::ShaderIncludeResolver::expandIncludes(assembledSource, currentDir,
-                                                                                  animIncludePaths, &includeError);
-        if (expanded.isEmpty()) {
-            qCWarning(lcEffect) << "Failed to expand shader includes for" << effectId << ":" << includeError;
-            return false;
-        }
-
-        // T1.1: splice the generated named-param preamble (`#define p_<id> ...`)
-        // after `#version`, identically to the daemon's loadFragmentShader. The
-        // accessors it emits (`customParams[N].xyzw`, `customColors[N]`,
-        // `uTexture<N>`) are declared in BOTH branches of animation_uniforms.glsl
-        // — including the PLASMAZONES_KWIN default-block branch injected below —
-        // so the same preamble compiles on this GL path and the daemon RHI path.
-        // Done before injectKwinDefineAfterVersion so the KWIN define still lands
-        // first after `#version`; the preamble's defines are pure text macros,
-        // expanded only where used (after the UBO include), so ordering is safe.
-        expanded = PhosphorShaders::spliceAfterVersion(
-            expanded, PhosphorAnimationShaders::AnimationShaderRegistry::paramPreamble(eff));
-
-        // HDR colour management. Installing this shader via
-        // OffscreenEffect::setShader REPLACES KWin's present shader, which
-        // normally carries the sRGB → output-colorspace conversion
-        // (ShaderTrait::TransformColorspace in KWin's own base.frag). Without
-        // it the transition writes sRGB verbatim into KWin's blending space —
-        // on an HDR output that space is gamma2.2 in the display's container
-        // colorimetry with 1.0 = peak luminance, so every window animation
-        // rendered dim and desaturated. Same bug class as the decoration
-        // present shader fix in decoration_render.cpp (kPresentFragment); the
-        // conversion below mirrors it, and KWin's base.frag, step for step:
-        // encoding → nits, colorimetry transform, tonemap, destination
-        // encoding.
-        //
-        // Deliberately NOT sourceEncodingToNitsInDestinationColorspace(): it
-        // folds doTonemapping() in at the end, which would tonemap twice — a
-        // double compression on HDR, precisely the case this exists for. No
-        // opacity modulate either: unlike the present shader, this path's
-        // window-rule opacity is folded into the sample by surfaceColor()
-        // (animation_uniforms.glsl), and this shader is compiled without the
-        // Modulate trait so KWin's TRAIT_MODULATE slot never applies.
-        //
-        // The override must be spliced AFTER expandIncludes ran above:
-        // colormanagement.glsl lives in KWin's `:/opengl/` resource, which
-        // KWin's GLShader::preprocess resolves for every source it compiles
-        // (custom ones included) — the phosphor include resolver would fail
-        // the whole expansion on it. The `#define` lands before the expanded
-        // animation_uniforms.glsl text, whose `#ifndef PZ_FINALIZE_COLOR`
-        // identity default then yields to this override; the generated entry
-        // main()s and bmw_compat's setOutputColor route their fragColor
-        // writes through the macro.
-        //
-        // The colorspace uniforms come free: KWin's OffscreenData::paint
-        // calls setColorspaceUniforms(sRGB, renderTarget.colorDescription(),
-        // Perceptual) on whatever shader setShader() installed. Do not push
-        // them from drawWindow — that is a redundant write KWin overwrites.
-        //
-        // On SDR this whole block is an exact no-op (identity colorimetry,
-        // round-tripping transfer functions, tonemap degrades to a clamp that
-        // cannot bite), so "SDR pixel-identical" is the regression test.
-        //
-        // The desktop-switch path (desktoptransitionmanager.cpp) deliberately
-        // does NOT get this splice: its capture FBOs inherit the output's
-        // colorDescription, so both blend inputs already live in the blending
-        // space and converting again would double-transform. It keeps the
-        // identity default.
-        static const QString kFinalizeColorBlock = QStringLiteral(
-            "#include \"colormanagement.glsl\"\n"
-            "vec4 pzFinalizeColor(vec4 c) {\n"
-            "    c = encodingToNits(c, sourceNamedTransferFunction,\n"
-            "                       sourceTransferFunctionParams.x, sourceTransferFunctionParams.y);\n"
-            "    c.rgb = (colorimetryTransform * vec4(c.rgb, 1.0)).rgb;\n"
-            "    c.rgb = doTonemapping(c.rgb);\n"
-            "    return nitsToDestinationEncoding(c);\n"
-            "}\n"
-            "#define PZ_FINALIZE_COLOR(c) pzFinalizeColor(c)\n");
-        expanded = PhosphorShaders::spliceAfterVersion(expanded, kFinalizeColorBlock);
-
-        // Selects the default-block branch in `animation_uniforms.glsl`.
-        // KWin's `KWin::GLShader` API addresses default-block uniforms only
-        // (no UBO bind path), so the canonical header's `#ifdef
-        // PLASMAZONES_KWIN` branch emits plain `uniform float iTime;`-style
-        // declarations instead of the daemon's `layout(std140, binding=0)
-        // uniform AnimationUniforms { ... };`. The macro must land AFTER the
-        // shader's `#version` line — the GLSL spec disallows tokens before
-        // `#version` other than whitespace and comments — so the helper
-        // below splices it between the version directive and the rest of
-        // the source.
-        const QByteArray fragWithKwinDefine = injectKwinDefineAfterVersion(expanded);
-
-        // Route the built-in default vertex source through the same injection
-        // as custom vertex stages so it gets the layout(location) extension
-        // enables (it uses explicit locations too, and KWin compiles it at
-        // #version 140 — see injectKwinDefineAfterVersion).
-        QByteArray vertWithKwinDefine = injectKwinDefineAfterVersion(QString::fromUtf8(kKwinDefaultVertexSource));
-        if (!eff.vertexShaderPath.isEmpty()) {
-            QFile vertFile(eff.vertexShaderPath);
-            if (!vertFile.open(QIODevice::ReadOnly)) {
-                qCWarning(lcEffect) << "Failed to open vertex shader" << eff.vertexShaderPath << "for effect"
-                                    << effectId << "— falling back to KWin default vertex stage";
-            } else {
-                const QString rawVert = QString::fromUtf8(vertFile.readAll());
-                if (rawVert.isEmpty()) {
-                    qCWarning(lcEffect) << "Vertex shader file is empty" << eff.vertexShaderPath << "for effect"
-                                        << effectId << "— falling back to KWin default vertex stage";
-                } else {
-                    const QString vertDir = QFileInfo(eff.vertexShaderPath).absolutePath();
-                    QString vertIncErr;
-                    const QString expandedVert = PhosphorShaders::ShaderIncludeResolver::expandIncludes(
-                        rawVert, vertDir, animIncludePaths, &vertIncErr);
-                    if (expandedVert.isEmpty()) {
-                        qCWarning(lcEffect) << "Failed to expand vertex shader includes for" << effectId << ":"
-                                            << vertIncErr << "— falling back to KWin default vertex stage";
-                    } else {
-                        // Same named-param preamble as the fragment stage: a
-                        // vertex-driven pack (wobble's velocity lag, the
-                        // pendulum swing) reads its `p_<id>` params in the
-                        // VERT, and the daemon bake already splices the
-                        // preamble into both stages — without this the GL
-                        // compile failed on the undefined identifiers and
-                        // the transition silently never installed.
-                        const QString vertWithParams = PhosphorShaders::spliceAfterVersion(
-                            expandedVert, PhosphorAnimationShaders::AnimationShaderRegistry::paramPreamble(eff));
-                        vertWithKwinDefine = injectKwinDefineAfterVersion(vertWithParams);
-                    }
-                }
-            }
-        }
-
-        // TransformColorspace is DECLARATIVE ONLY here, exactly as on the
-        // decoration present shader (decoration_render.cpp): with a custom
-        // fragment source generateCustomShader uses the traits for nothing
-        // but listDefines, and colormanagement.glsl never reads
-        // TRAIT_TRANSFORM_COLORSPACE. The conversion is the explicit
-        // pzFinalizeColor splice above and runs with or without this flag; it
-        // is declared so the shader's traits describe what it actually does.
-        auto shader = KWin::ShaderManager::instance()->generateCustomShader(
-            KWin::ShaderTrait::MapTexture | KWin::ShaderTrait::TransformColorspace, vertWithKwinDefine,
-            fragWithKwinDefine);
-        // KWin 6.7 removed GLShader::isValid(); generateCustomShader now returns
-        // nullptr when compilation or linking fails, so a null check is the
-        // validity test.
-        if (!shader) {
-            qCWarning(lcEffect) << "Failed to compile shader transition" << effectId
-                                << "— caching the failure so subsequent transitions skip the recompile "
-                                   "until the next shader hot-reload.";
-            // Cache a null-shader sentinel. A failed compile must NOT be
-            // retried on every transition: without this the cache miss recurs
-            // each time and the full read+assemble+expand+compile re-runs on
-            // the compositor thread — a per-command stall (the same failure
-            // mode the GLSL #extension fix addressed for the morph shader).
-            // The sentinel is distinguishable from a live entry because a
-            // successful compile always emplaces a non-null shader. It is
-            // cleared by the effectsChanged handler's m_shaderCache.clear()
-            // (shader hot-reload / settings change), so a corrected shader
-            // recompiles on the next reload.
-            m_shaderManager.m_shaderCache.emplace(effectId, CachedShader{});
-            return false;
-        }
-
-        CachedShader cached;
-        // Animation-shader contract — names sourced from
-        // `PhosphorAnimationShaders::AnimationShaderContract`. Both the
-        // daemon overlay-surface execution site and this compositor
-        // window-content execution site resolve the same names.
-        cached.iTimeLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kITime);
-        cached.iResolutionLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIResolution);
-        cached.iTimeDeltaLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kITimeDelta);
-        cached.iFrameLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIFrame);
-        cached.iDateLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIDate);
-        cached.iMouseLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIMouse);
-        cached.iIsReversedLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIIsReversed);
-        cached.iSurfaceScreenPosLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kISurfaceScreenPos);
-        cached.iAnchorSizeLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIAnchorSize);
-        cached.iAnchorPosInFboLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIAnchorPosInFbo);
-        cached.iAnchorRectInTextureLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIAnchorRectInTexture);
-        // Geometry-morph uniforms (snap / maximize / layout-switch
-        // cross-fade). -1 when the shader is not a morph shader (doesn't
-        // read them) — paintWindow guards on >= 0 so non-morph transitions
-        // pay nothing.
-        cached.iFromRectLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIFromRect);
-        cached.iToRectLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIToRect);
-        // Task-manager icon rect for minimize-to-icon packs (genie,
-        // phosphor-siphon). -1 for every pack that doesn't declare it —
-        // same pay-nothing guard.
-        cached.iIconRectLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIIconRect);
-        cached.iOldWindowLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kUOldWindow);
-        // Surface-layer-stack uniforms — every animation shader resolves these
-        // (declared in the shared header and read through surfaceColor()), so
-        // they are valid whenever the shader samples the window surface. The
-        // kwin-effect binds the layered surface + flag each frame; a window with
-        // no surface layers pushes the flag as 0 and the shader samples the bare
-        // uTexture0. See AnimationShaderContract::kUSurfaceLayer / kIHasSurfaceLayer.
-        cached.uSurfaceLayerLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kUSurfaceLayer);
-        cached.iHasSurfaceLayerLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIHasSurfaceLayer);
-        cached.iHasOldWindowLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIHasOldWindow);
-        // Audio spectrum (opt-in audio.glsl module; -1 for non-audio packs).
-        cached.iAudioSpectrumSizeLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIAudioSpectrumSize);
-        cached.uAudioSpectrumLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kUAudioSpectrum);
-        cached.iMoveVelocityLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIMoveVelocity);
-        cached.iMoveOffsetLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIMoveOffset);
-        cached.iMoveVelocity2Loc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIMoveVelocity2);
-        cached.iMoveTrailLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIMoveTrail);
-        cached.iMoveMeshLoc = shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIMoveMesh);
-        cached.iLayerRectInTextureLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kILayerRectInTexture);
-        // uTexture0 — for the transition-time composite retarget (see
-        // CachedShader::uTexture0Loc). The name is the contract's literal
-        // sampler identifier (animation_uniforms.glsl, both branches).
-        cached.uTexture0Loc = shader->uniformLocation("uTexture0");
-        // SetOpacity rule opacity — a separate concern from the morph uniforms
-        // above: applies to ALL shaders (compositor path only), so surfaceColor
-        // can dim the surface for a SetOpacity rule. See
-        // AnimationShaderContract::kIWindowOpacity.
-        cached.iWindowOpacityLoc =
-            shader->uniformLocation(PhosphorAnimationShaders::AnimationShaderContract::kIWindowOpacity);
-        // Cache element locations for the per-effect declared parameter
-        // slots: `customParams[0..kMaxCustomParams-1]` for float / int /
-        // bool params, and `customColors[0..kMaxCustomColors-1]` for color
-        // params. Each declared parameter lands in one of these slots —
-        // see `AnimationShaderRegistry::translateAnimationParams` for
-        // the exact mapping. `glGetUniformLocation` returns -1 for slots
-        // the shader didn't reference (e.g. a one-param effect that the
-        // GLSL compiler optimises away the unused tail of either array);
-        // the per-frame push loop in paintWindow guards against -1 to
-        // skip the setUniform call.
-        for (int slot = 0; slot < PhosphorAnimationShaders::AnimationShaderContract::kMaxCustomParams; ++slot) {
-            // Pre-baked element-name table — no per-slot QByteArray
-            // alloc. Sized + static_asserted against the contract budget
-            // at the namespace-level definition.
-            cached.customParamsLoc[slot] = shader->uniformLocation(kCustomParamsElementNames[slot]);
-        }
-        for (int slot = 0; slot < PhosphorAnimationShaders::AnimationShaderContract::kMaxCustomColors; ++slot) {
-            cached.customColorsLoc[slot] = shader->uniformLocation(kCustomColorsElementNames[slot]);
-        }
-        // User textures: resolve sampler + iTextureResolution[N] uniform
-        // locations only. The actual texture upload happens per-leg
-        // inside `beginShaderTransition`'s body below — keyed by the
-        // resolved path in `m_shaderManager.m_textureCache` so two legs with different
-        // override paths don't collide on the per-effect cache.
-        for (int slot = 0; slot < PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots; ++slot) {
-            // GLSL sampler name: uTexture1..3 (slot+1 because uTexture0 is
-            // the redirected surface, not user-declared). Matches the
-            // overlay shader convention in data/overlays/shared/textures.glsl.
-            // Pre-baked from the file-scope `kUserTextureSamplerNames` /
-            // `kITextureResolutionKeys` arrays — no per-slot QByteArray
-            // alloc per shader install.
-            cached.userTextureLoc[slot] = shader->uniformLocation(kUserTextureSamplerNames[slot]);
-            cached.iTextureResolutionLoc[slot] = shader->uniformLocation(kITextureResolutionKeys[slot]);
-        }
-        cached.shader = std::move(shader);
-        cacheIt = m_shaderManager.m_shaderCache.emplace(effectId, std::move(cached)).first;
+    // Compile (or fetch the cached) shader program + uniform locations for this
+    // effect. Split out to shader_textures.cpp; the GL context made current just
+    // above is the one it compiles against.
+    const CachedShader* cachedShaderEntry = compileOrLoadAnimationShader(effectId, eff);
+    if (!cachedShaderEntry) {
+        // Transient failure (missing / empty / unexpandable source) that
+        // compileOrLoadAnimationShader deliberately does NOT cache — install
+        // nothing and let a later trigger re-attempt.
+        return false;
     }
-
     // A cached null-shader sentinel marks a prior compile failure (see the
-    // "Failed to compile" branch above). Skip the transition without
-    // re-attempting the expensive compile on every trigger; the effectsChanged
-    // handler clears m_shaderCache on hot-reload, so a corrected shader
-    // recompiles then. A successfully compiled entry always holds a non-null
-    // shader, so this never false-positives on a live transition.
-    if (!cacheIt->second.shader) {
+    // "Failed to compile" branch in compileOrLoadAnimationShader). Skip the
+    // transition without re-attempting the expensive compile on every trigger;
+    // the effectsChanged handler clears m_shaderCache on hot-reload, so a
+    // corrected shader recompiles then. A successfully compiled entry always
+    // holds a non-null shader, so this never false-positives on a live transition.
+    if (!cachedShaderEntry->shader) {
         return false;
     }
 
@@ -1100,7 +412,7 @@ bool PlasmaZonesEffect::beginShaderTransition(KWin::EffectWindow* window,
     // (durationMs > 0) also match — a true "second trigger of the same
     // already-playing transition" — so a reverse leg or a mode flip
     // still supersedes correctly.
-    if (isSameWindowSupersession && existing->cached == &cacheIt->second && existing->reverse == reverse
+    if (isSameWindowSupersession && existing->cached == cachedShaderEntry && existing->reverse == reverse
         && ((existing->durationMs > 0) == (durationMs > 0))) {
         // Same-effect short-circuit: prior leg is intact and continues
         // running. Caller (`tryBeginShaderForEvent`) MUST NOT schedule a
@@ -1150,7 +462,7 @@ bool PlasmaZonesEffect::beginShaderTransition(KWin::EffectWindow* window,
     // redirect() call below (no-op endShaderTransition since the map
     // doesn't have the entry).
 
-    const auto& cachedEntry = cacheIt->second;
+    const auto& cachedEntry = *cachedShaderEntry;
     ShaderTransition transition;
     transition.cached = &cachedEntry;
     // Surface-extent shaders (metadata `fboExtent: "surface"`) render past
@@ -1826,597 +1138,6 @@ void PlasmaZonesEffect::endShaderTransition(KWin::EffectWindow* window)
             },
             Qt::QueuedConnection);
     }
-}
-
-bool PlasmaZonesEffect::resolvedShaderAppliesToEvent(const QString& effectId, const QString& profilePath) const
-{
-    // See the header doc. Routed through the canonical predicate
-    // (shaderEffectAppliesToEventPath) — the same one the settings pickers
-    // filter with — so runtime refusal and picker filtering can never drift.
-    const auto eff = m_shaderManager.m_animationShaderRegistry.effect(effectId);
-    if (!eff.isValid()) {
-        // Unknown id: pass through. The pack may still be scanning, and
-        // beginShaderTransition's registry-miss warning stays the single
-        // reporter for genuinely unknown ids.
-        return true;
-    }
-    if (!PhosphorAnimationShaders::shaderEffectAppliesToEventPath(eff, profilePath)) {
-        qCDebug(lcEffect) << "shader" << effectId << "does not apply to event" << profilePath
-                          << "(appliesTo=" << eff.appliesTo << ") — skipping transition";
-        return false;
-    }
-    return true;
-}
-
-void PlasmaZonesEffect::tryBeginShaderForEvent(KWin::EffectWindow* window, const QString& profilePath, int durationMs,
-                                               bool reverse, bool holdCloseGrab, bool holdAddedGrab,
-                                               bool animateMinimized)
-{
-    if (!window || durationMs <= 0) {
-        // Defensive guard. The current call sites all pass
-        // `animationDurationMs()` which the daemon-bringup loader
-        // clamps to `[MinAnimationDurationMs, MaxAnimationDurationMs]`
-        // = [50, 2000], so 0 cannot reach this code through normal
-        // flow. The authoritative no-animations gate is
-        // `m_windowAnimator->isEnabled()` checked just below — that
-        // covers the user-toggled case. This guard exists to fail
-        // closed if a future programmatic call site bypasses the
-        // clamp; a Timing Rule intentionally cannot rescue a
-        // 0/negative duration since the value is treated as "caller
-        // didn't supply one" rather than the "inherit per-event
-        // default" sentinel that the rule layer recognises.
-        return;
-    }
-    // Fast-path early-out on the global animations toggle. The
-    // authoritative gate also lives in `beginShaderTransition` (so
-    // window.movement.* callers via `applyWindowGeometry` are gated too), but
-    // dispatching there would still pay the shader-tree resolve cost
-    // — this skips it entirely when the global toggle is off.
-    if (!m_windowAnimator->isEnabled()) {
-        return;
-    }
-    // Window-filtering gate. `shouldAnimateWindow` honours the user's
-    // Animations.WindowFiltering exclusions (transient / min-size /
-    // app / class) AND lets a Rule carrying any effect-consumed
-    // (Tag::Effect) action override the filter when the rule's match
-    // expression resolves for the window's full WindowQuery (AppId /
-    // WindowClass / Title / WindowRole / DesktopFile / WindowType / Pid /
-    // state flags). Skipping this for shader transitions only would leave
-    // the motion-side cascade in `applyWindowGeometry` doing its own check;
-    // both call sites gate identically so the filter is a single concept
-    // across the two paths.
-    if (!shouldAnimateWindow(window)) {
-        return;
-    }
-    // Cascade: per-window animation Rule → ShaderProfileTree
-    // (per-event default). The rule layer wins for matching windows;
-    // an engaged-empty effectId on the rule deliberately blocks the
-    // tree fallthrough (the user's "no animation for this app on this
-    // event" sentinel).
-    //
-    // Build the full per-window query once and reuse it for every
-    // resolver call below — same shape `shouldAnimateWindow` already
-    // uses for the rule-override gate, so a rule that passes the gate
-    // also resolves its slot. Caching across resolver calls is built
-    // into the evaluator's `resolveCached(windowId, …)` path; the query
-    // here is only the match input, not the cache key.
-    const PhosphorRules::WindowQuery query = ruleQuery(window);
-    const QString windowId = getWindowId(window);
-    const auto& profileTree = m_shaderManager.profileTree();
-    // Per-event motion profile (curve + duration) in ONE walk, via the shared
-    // SSOT: global animator profile → category "All" → per-node motion-tree
-    // override → per-window Rule. The daemon mirrors its motion
-    // PhosphorProfileRegistry into `motionProfileTree` over D-Bus, so a user-set
-    // `window.open` = 900 ms (or an "All" curve) wins over the global default.
-    //
-    // The base is the animator's global profile, so when NO node overrides the
-    // result is the global (never the library 150 ms default), and BOTH the
-    // duration and the curve come from the SAME base — no cross-field base
-    // skew. `effectiveDuration()` feeds the combined resolver as its
-    // `defaultDurationMs` (the Rule timing slot still layers on top, matching
-    // the resolver's documented rule → per-event → global contract), and
-    // `.curve` shapes the time-driven `iTime`: paintWindow eases the linear
-    // progress through it so a node's curve (e.g. "Ease Out") applies to its
-    // shader exactly as it does on the animator-driven snap path. Null curve →
-    // linear iTime.
-    const PhosphorAnimation::Profile eventMotion = resolveEventMotionProfile(profilePath, query, windowId);
-    const int baseDurationMs = qRound(eventMotion.effectiveDuration());
-    const std::shared_ptr<const PhosphorAnimation::Curve> progressCurve = eventMotion.curve;
-    // Combined cascade: ONE cached evaluator walk feeds BOTH the shader-slot
-    // and timing-slot reads. The pre-refactor pair of `resolveAnimationShader
-    // Profile` + `resolveAnimationDuration` ran two priority-order walks per
-    // event (same query, both bypassing the per-window match cache); the
-    // combined shim issues a single `resolveCached(windowId, …)` and reads
-    // both slots from the same `ResolvedActions`. Semantics are identical:
-    // rule wins per-slot, with engaged-empty effectId still blocking the tree
-    // fallthrough and durationMs <= 0 still meaning "inherit".
-    //
-    // Clamp the resolved duration to the upstream `durationMs` floor: if
-    // the cascade collapses to <= 0 (corrupt persisted rule, missing
-    // motion-tree node feeding baseDurationMs), the QTimer::singleShot
-    // below would fire on the next event-loop tick and tear down the
-    // just-installed transition before its first paint. The input
-    // `durationMs` was already clamped by the daemon-bringup loader to
-    // [MinAnimationDurationMs, MaxAnimationDurationMs], and the
-    // `durationMs <= 0` guard at the top of `tryBeginShaderForEvent`
-    // rejects non-positive inputs, so `durationMs` here is a safe
-    // positive floor.
-    const auto resolved = PlasmaZones::resolveAnimationShaderProfile(m_shaderManager.animationRuleEvaluator(),
-                                                                     profileTree, windowId, query, profilePath);
-    const auto& profile = resolved.profile;
-    // The duration comes from the motion cascade ALONE. resolveEventMotionProfile
-    // already applied the Rule timing slot and clamped the result into the
-    // envelope, so there is exactly one read and one clamp of that slot; the shader
-    // resolver deliberately no longer re-reads it.
-    int effectiveDurationMs = baseDurationMs;
-    if (effectiveDurationMs <= 0) {
-        effectiveDurationMs = durationMs;
-    }
-    // Spring lifetime, shared with the desktop switch: a stateful curve derives its
-    // own lifetime from settleTime() and ignores the duration entirely. The result
-    // drives BOTH the paint active-window and the teardown timer below, so the two
-    // stay in lockstep.
-    effectiveDurationMs = ShaderInternal::resolveTransitionLifetimeMs(effectiveDurationMs, progressCurve.get());
-    if (profile.effectiveEffectId().isEmpty()) {
-        // Default-state path: a fresh user with no shader overrides
-        // anywhere in the tree resolves every event to empty effectId,
-        // which is correct ("no shader assigned"). Logging at WARNING
-        // for that floods the journal with bogus failures every time a
-        // window opens, closes, or moves. Only WARN when the tree has
-        // overrides (so an empty resolve here is genuinely surprising —
-        // the documented prune / D-Bus-race scenarios), otherwise
-        // demote to DEBUG.
-        const int ruleCount = m_shaderManager.animationRuleSet().count();
-        if (profileTree.overriddenPaths().isEmpty() && ruleCount == 0) {
-            qCDebug(lcEffect) << "tryBeginShader[" << profilePath
-                              << "]: no shader assigned (tree empty — default state)";
-        } else {
-            qCWarning(lcEffect) << "tryBeginShader[" << profilePath
-                                << "]: no shader assigned (cascade returned empty effectId, tree size="
-                                << profileTree.overriddenPaths().size() << " rules=" << ruleCount << ")";
-        }
-        return;
-    }
-    // Runtime applicability gate — see resolvedShaderAppliesToEvent. The
-    // pickers keep class-mismatched packs unselectable, but a Rule's
-    // OverrideAnimationShader slot or a stale / hand-edited config bypasses
-    // them. Most material on the held-drag leg (window.movement.move): a
-    // crossfade pack there would install a dead transition that pins
-    // full-output repaints for the whole drag. The tree itself can no longer
-    // deliver one there (ShaderProfileTree::resolve takes no ancestor overlay
-    // for the move leaf), so this catches the rule-layer and stale-config
-    // routes — and, symmetrically, a move-physics or desktop pack forced onto
-    // any other window leg.
-    if (!resolvedShaderAppliesToEvent(profile.effectiveEffectId(), profilePath)) {
-        return;
-    }
-    const bool installed = beginShaderTransition(window, profile, effectiveDurationMs, reverse, holdCloseGrab,
-                                                 holdAddedGrab, animateMinimized, progressCurve);
-    auto* transition = m_shaderManager.findTransition(window);
-    // Mark the held-move leg by IDENTITY. The drag handlers must not infer it from
-    // liveness: `window.movement.move` is opt-in with no default shader, so the
-    // common case installs nothing at all and `findTransition` would hand them an
-    // unrelated leg to pin and reverse. See ShaderTransition::heldMove.
-    //
-    // We may stamp when `installed` is false, but ONLY for the same-effect
-    // short-circuit — where the live leg genuinely IS the pack this event resolved
-    // (a pack whose `appliesTo` admits both "move" and another class can already be
-    // running for that other event). `beginShaderTransition` returns false from many
-    // other places — compile failure, the cached null-shader sentinel, a registry
-    // miss, a refused pack, a collapsed surface — and in EVERY one of those the live
-    // leg is something else entirely, most reachably the `window.focus` leg the click
-    // that began this drag just installed. Stamping that would pin it for the drag,
-    // kill its teardown timer, and play the focus animation BACKWARD on release: the
-    // exact bug this flag exists to prevent, re-introduced from the write side.
-    //
-    // So test what the short-circuit itself tests — does the live leg's cached shader
-    // point at THIS event's pack? The null-shader sentinel is excluded by the
-    // `->shader` check, which matters because that sentinel is sticky: once a pack
-    // fails to compile, every later drag in the session would otherwise mis-stamp.
-    //
-    // Only ever write TRUE. A non-move event short-circuiting into a live held-move
-    // leg must leave the flag alone — supersession builds a fresh ShaderTransition
-    // whose default is already false, so the false case needs no code, and writing it
-    // would re-introduce the mislabelling from the other direction.
-    bool ownsResolvedLeg = installed;
-    if (!ownsResolvedLeg && transition) {
-        const auto cacheIt = m_shaderManager.m_shaderCache.find(profile.effectiveEffectId());
-        ownsResolvedLeg = cacheIt != m_shaderManager.m_shaderCache.end() && cacheIt->second.shader
-            && transition->cached == &cacheIt->second;
-    }
-    if (transition && ownsResolvedLeg && profilePath == PhosphorAnimation::ProfilePaths::WindowMove) {
-        transition->heldMove = true;
-    }
-    if (!installed || !transition) {
-        // Either beginShaderTransition no-op'd (compile fail, invalid id,
-        // collapsed surface, animations disabled) and there is nothing
-        // to teardown, OR the same-effect short-circuit kept the prior
-        // leg in flight — in which case the prior leg's own teardown
-        // timer (or animator-completion callback) owns the teardown.
-        // Scheduling a fresh timer here would carry the prior leg's
-        // generation and fire on this event's (likely shorter) duration,
-        // cutting the still-running animation short.
-        return;
-    }
-    // Capture the just-installed transition's generation so the deferred
-    // teardown bails if a successor has replaced us by the time the timer
-    // fires. Without this, two events overlapping on the same window
-    // (window.move during window.snapIn, window.focus interrupting
-    // window.maximize) leave a stale timer that tears down the SUCCESSOR
-    // when its own timer hasn't fired yet.
-    scheduleShaderTransitionTeardown(window, transition->generation, effectiveDurationMs);
-}
-
-void PlasmaZonesEffect::scheduleShaderTransitionTeardown(KWin::EffectWindow* window, quint64 generation, int delayMs)
-{
-    QPointer<KWin::EffectWindow> safeWindow(window);
-    QTimer::singleShot(qMax(1, delayMs), this, [this, safeWindow, generation]() {
-        // Two-tier guard: QPointer catches QObject destruction,
-        // endShaderTransition's isDeleted() catches KWin's deletion-animation phase
-        if (!safeWindow) {
-            return;
-        }
-        const auto* live = m_shaderManager.findTransition(safeWindow);
-        if (!live || live->generation != generation) {
-            // A newer transition replaced us (last-event-wins) and owns its own
-            // timer — leave it alone.
-            return;
-        }
-        // HELD transitions (the interactive drag) outlive their nominal duration by
-        // design: the user is still dragging. windowFinishUserMovedResized owns
-        // their teardown (a settle tail after release), so the duration timer stands
-        // down. A mesh-backed drag released BEFORE this timer fires is covered by the
-        // generation check above instead: the release handler clears the hold flag
-        // but bumps the generation when it hands the lifetime to the settle gate.
-        if (live->holdUntilRelease) {
-            return;
-        }
-        // Re-check against the transition's OWN clock rather than trusting the delay
-        // we were armed with. `startTimeMs` is REBASED every frame a window spends
-        // under restore suppression (a window repositioned on open is withheld from
-        // compositing until its configure lands, and its animation must not play
-        // invisibly in the meantime — see paint_pipeline). The install-time arming is
-        // therefore up to kRestoreSuppressDeadlineMs (250 ms) too early, and firing
-        // it would tear the leg down while its timeline still had a quarter second
-        // to run — the open animation is cut mid-flight and the window pops. Re-arm
-        // for the remainder instead. Any future rebase gets the same treatment for
-        // free, which is why this is a re-check and not a suppression special case.
-        const qint64 remaining =
-            static_cast<qint64>(live->durationMs) - (ShaderInternal::shaderClockNowMs() - live->startTimeMs);
-        if (remaining > 0) {
-            scheduleShaderTransitionTeardown(safeWindow.data(), generation,
-                                             static_cast<int>(qMin<qint64>(remaining, 60000)));
-            return;
-        }
-        endShaderTransition(safeWindow);
-    });
-}
-
-void PlasmaZonesEffect::loadShaderProfileFromDbus()
-{
-    // The key is named INLINE, not bound to a local alias first. An alias saves nothing
-    // here and costs real safety: the registry-contract test scrapes these call sites to
-    // prove every key the effect fetches is registered daemon-side, and an alias forces
-    // it to resolve an identifier back to a constant — which it has now got wrong three
-    // separate times, each time silently checking the wrong key while its own self-check
-    // balanced. Name the constant where it is used and there is nothing to resolve.
-    PhosphorProtocol::ClientHelpers::loadSettingAsync(
-        this, PhosphorProtocol::Service::SettingProperty::ShaderProfileTree, [this](const QVariant& v) {
-            dispatchJsonSetting(PhosphorProtocol::Service::SettingProperty::ShaderProfileTree, v,
-                                [this](const QJsonObject& obj) {
-                                    auto& tree = m_shaderManager.profileTree();
-                                    tree = PhosphorAnimationShaders::ShaderProfileTree::fromJson(obj);
-                                    qCDebug(lcEffect) << "loadShaderProfileFromDbus: tree loaded with"
-                                                      << tree.overriddenPaths().size()
-                                                      << "overrides — paths=" << tree.overriddenPaths();
-                                    // A tree edit can assign or unassign an audio-reactive
-                                    // animation pack; re-evaluate the cava run gate so the
-                                    // provider is warm before the first transition needs it.
-                                    scheduleEffectAudioSync();
-                                    // It can also assign or clear any suppression-owning pack
-                                    // (`desktop.peek`, `window.minimize`, `window.maximize`);
-                                    // keep KWin's own stock effects unloaded exactly while
-                                    // ours owns the event.
-                                    syncStockEffectSuppression();
-                                },
-                                /*arraySink=*/{});
-        });
-}
-
-void PlasmaZonesEffect::slotRulesChanged()
-{
-    // Coalesce burst signals: the daemon emits one `rulesChanged` per per-rule
-    // mutation, so a 50-rule batch edit would otherwise drive 50 sequential
-    // `getAllRules` round-trips + JSON parses + filter walks. The timer is a
-    // single-shot 50ms debounce (set up in the constructor); each call here
-    // re-arms it, so only the trailing edge of the burst triggers a refresh.
-    m_animationRulesRefreshDebounce.start();
-}
-
-void PlasmaZonesEffect::loadRuleAnimationsFromDbus()
-{
-    // Fetch the unified Rule store via getAllRules (returns a JSON
-    // string of a v4 RuleSet), deserialise, filter to rules whose
-    // action list contains any effect-consumed (Tag::Effect) action, and
-    // hand them to the shader manager. The shader manager mirrors them into
-    // m_animationRuleSet so the per-event slot lookup in shader_resolve.cpp
-    // resolves the cascade against the unified rule store directly.
-    const QDBusMessage msg = QDBusMessage::createMethodCall(
-        QString(PhosphorProtocol::Service::Name), QString(PhosphorProtocol::Service::ObjectPath),
-        QString(PhosphorProtocol::Service::Interface::Rules), QStringLiteral("getAllRules"));
-    const QDBusPendingCall pending = QDBusConnection::sessionBus().asyncCall(msg);
-    auto* watcher = new QDBusPendingCallWatcher(pending, this);
-    connect(watcher, &QDBusPendingCallWatcher::finished, this, [this](QDBusPendingCallWatcher* w) {
-        w->deleteLater();
-        const QDBusPendingReply<QString> reply = *w;
-        if (reply.isError()) {
-            // Daemon may not be up yet at startup; the rulesChanged
-            // subscription below will deliver the next change. Log at debug
-            // so the noise stays out of normal-startup logs.
-            qCDebug(lcEffect) << "loadRuleAnimationsFromDbus: getAllRules failed:" << reply.error().message();
-            return;
-        }
-        const QByteArray payload = reply.value().toUtf8();
-        const QJsonDocument doc = QJsonDocument::fromJson(payload);
-        if (!doc.isObject()) {
-            qCWarning(lcEffect) << "loadRuleAnimationsFromDbus: getAllRules returned non-object JSON";
-            return;
-        }
-        const auto setOpt = PhosphorRules::RuleSet::fromJson(doc.object());
-        if (!setOpt) {
-            qCWarning(lcEffect) << "loadRuleAnimationsFromDbus: RuleSet::fromJson refused payload";
-            return;
-        }
-        // Sample the prior rule set for SetOpacity BEFORE setRuleAnimationRules
-        // overwrites it. Repaint is needed on BOTH bookends — rule appears
-        // (currently-natural-opacity windows need to apply it) AND rule
-        // disappears (currently-dimmed windows need to revert). The earlier
-        // single-bookend form left previously-dimmed windows stuck at their
-        // last-painted opacity when the user removed the last SetOpacity rule.
-        bool hadSetOpacity = false;
-        const auto& priorRules = m_shaderManager.animationRuleSet().rules();
-        for (const PhosphorRules::Rule& rule : priorRules) {
-            for (const PhosphorRules::RuleAction& action : rule.actions) {
-                if (action.type == PhosphorRules::ActionType::SetOpacity) {
-                    hadSetOpacity = true;
-                    break;
-                }
-            }
-            if (hadSetOpacity) {
-                break;
-            }
-        }
-
-        QList<PhosphorRules::Rule> animationRules;
-        for (const PhosphorRules::Rule& rule : setOpt->rules()) {
-            if (!rule.enabled) {
-                // Skip disabled rules — they exist in the store but must not
-                // contribute to the evaluator. (RuleEvaluator already gates
-                // on enabled for its own walks, but pruning here keeps the
-                // rule-set size minimal and the priority-order index smaller.)
-                continue;
-            }
-            // Admit the rule to the evaluator if ANY action is effect-consumed,
-            // i.e. carries Tag::Effect (hasTag below). The authoritative
-            // membership list is the descriptor tag assignments in
-            // ruleaction.cpp — animation overrides, SetOpacity, the appearance
-            // family (SetBorder*, SetHideTitleBar, OverrideDecorationChain),
-            // and SetWindowLayer.
-            bool admitted = false;
-            for (const PhosphorRules::RuleAction& action : rule.actions) {
-                if (PhosphorRules::ActionRegistry::instance().hasTag(action.type, PhosphorRules::Tag::Effect)) {
-                    admitted = true;
-                    break;
-                }
-            }
-            if (admitted) {
-                animationRules.append(rule);
-            }
-        }
-        m_shaderManager.setRuleAnimationRules(std::move(animationRules));
-        // A rule edit can route transitions to (or away from) an audio-reactive
-        // animation pack via an EffectId payload — re-evaluate the cava run gate.
-        scheduleEffectAudioSync();
-        // The new-state SetOpacity predicate is computed by rebuildAnimationRuleSet
-        // (see ShaderTransitionManager::hasOpacityRules) — read it back rather than
-        // re-scanning the rule list a second time here.
-        const bool hasSetOpacity = m_shaderManager.hasOpacityRules();
-        qCDebug(lcEffect) << "loadRuleAnimationsFromDbus: forwarded" << m_shaderManager.animationRuleSet().count()
-                          << "total animation rules to the evaluator";
-
-        // Per-window border / title-bar rules ride the same animation rule set
-        // (Tag::Effect admits them). Refresh borders so an edited /
-        // added / removed SetBorder* / SetHideTitleBar rule applies immediately
-        // — updateAllDecorations re-merges every window and reconciles rule-hidden
-        // title bars against the fresh evaluator.
-        updateAllDecorations();
-
-        // Update the drag-gate exclusion rule set from the same unified
-        // payload — `loadRuleAnimationsFromDbus` is the effect's one
-        // and only rule-store sync point, so the snapping-exclusion gate
-        // refreshes here too rather than chasing a second D-Bus fetch. The
-        // filter keeps only enabled rules with a terminal Exclude action;
-        // setRules bumps the bound rule set's revision so
-        // m_snappingExclusionEvaluator's per-revision sort index rebuilds
-        // on its next walk (these evaluators call uncached `resolve()`, so
-        // there is no per-window match cache to drop — the sort index is
-        // the only revision-keyed artifact).
-        m_snappingExclusionRuleSet.setRules(PhosphorRules::ExclusionRules::excludeRulesFrom(*setOpt).rules());
-
-        // Same refresh for the animation-side exclusion rule set, sliced
-        // for `ExcludeAnimations`-action rules. The two slices stay
-        // independent so a user can have a window excluded from animations
-        // but NOT from snap (or vice versa).
-        m_animationExclusionRuleSet.setRules(
-            PhosphorRules::ExclusionRules::excludeAnimationsRulesFrom(*setOpt).rules());
-        // Force a full repaint on EITHER bookend so a user-authored rule
-        // applies to static (un-damaged) windows immediately AND so a
-        // removed rule reverts previously-dimmed windows immediately, not
-        // on the next incidental damage event. OverrideAnimation* rules
-        // fire on lifecycle events so they don't need this kick;
-        // SetOpacity continuously alters paint output regardless of
-        // animation state and needs the kick on both transitions.
-        if ((hasSetOpacity || hadSetOpacity) && KWin::effects) {
-            KWin::effects->addRepaintFull();
-        }
-    });
-}
-
-void PlasmaZonesEffect::loadMotionProfileTreeFromDbus()
-{
-    // The key is named INLINE, not bound to a local alias first. An alias saves nothing
-    // here and costs real safety: the registry-contract test scrapes these call sites to
-    // prove every key the effect fetches is registered daemon-side, and an alias forces
-    // it to resolve an identifier back to a constant — which it has now got wrong three
-    // separate times, each time silently checking the wrong key while its own self-check
-    // balanced. Name the constant where it is used and there is nothing to resolve.
-    PhosphorProtocol::ClientHelpers::loadSettingAsync(
-        this, PhosphorProtocol::Service::SettingProperty::MotionProfileTree, [this](const QVariant& v) {
-            dispatchJsonSetting(PhosphorProtocol::Service::SettingProperty::MotionProfileTree, v,
-                                [this](const QJsonObject& obj) {
-                                    // ProfileTree::fromJson resolves each node's optional
-                                    // `curve` field through a CurveRegistry. The effect now
-                                    // resolves per-event curve AND duration from this tree
-                                    // (resolveEventMotionProfile reads `.curve` to shape the
-                                    // time-driven iTime), so parse with the effect's own
-                                    // `m_curveRegistry` — the SAME registry the Rule path
-                                    // resolves against (shader_resolve.cpp) — rather than a
-                                    // throwaway.
-                                    //
-                                    // Builtins are sufficient here and no CurveLoader is
-                                    // needed in the compositor: a Profile persists its curve
-                                    // BY SPEC, not by registry key (Profile::toJson writes
-                                    // `curve->toString()`, e.g. "0.34,1.20,0.64,1.00" or
-                                    // "spring:<omega>,<zeta>"; the friendly name rides in the
-                                    // separate `presetName` field). A user curve pack
-                                    // (data/curves/*.json) is a named preset over a BUILTIN
-                                    // typeId, so it round-trips through its spec and the
-                                    // ctor-registered builtin factories resolve it. Sharing
-                                    // one registry keeps both curve paths on identical
-                                    // resolution rules rather than two that can drift.
-                                    auto& tree = m_shaderManager.motionProfileTree();
-                                    tree = PhosphorAnimation::ProfileTree::fromJson(obj, m_curveRegistry);
-                                    qCDebug(lcEffect) << "loadMotionProfileTreeFromDbus: tree loaded with"
-                                                      << tree.overriddenPaths().size()
-                                                      << "per-event overrides — paths=" << tree.overriddenPaths();
-                                },
-                                /*arraySink=*/{});
-        });
-}
-
-PhosphorAnimation::Profile PlasmaZonesEffect::resolveEventMotionProfile(const QString& profilePath,
-                                                                        const PhosphorRules::WindowQuery& query,
-                                                                        const QString& windowId) const
-{
-    // Cascade base: the WindowAnimator's global profile carries the authoritative
-    // global curve + duration (from animationEasingCurve / animationDuration).
-    //
-    // Before the async settings load lands, the animator's `duration` is still
-    // nullopt, so effectiveDuration() falls back to Profile::DefaultDuration
-    // while `animationDurationMs()` reports Limits::DefaultAnimationDurationMs.
-    // Callers rely on those two agreeing (it is what makes the pre-load window
-    // resolve to the same duration either way), so pin the coupling here rather
-    // than let a future bump to one silently skew it.
-    static_assert(
-        qRound(PhosphorAnimation::Profile::DefaultDuration) == PhosphorAnimation::Limits::DefaultAnimationDurationMs,
-        "Profile::DefaultDuration and Limits::DefaultAnimationDurationMs must agree, or the pre-settings-load "
-        "motion cascade skews against animationDurationMs()");
-    const PhosphorAnimation::Profile& base = m_windowAnimator->profile();
-    // Category "All" → per-node: overlay only the motion-tree override chain for
-    // this path onto the global base (overlayChainOnto skips the tree's own
-    // baseline and returns base untouched when nothing in the chain overrides,
-    // so an empty tree keeps the animator's global). Gated on a non-empty
-    // override set to keep the default-state fast-path (no chain walk).
-    const auto& motionTree = m_shaderManager.motionProfileTree();
-    PhosphorAnimation::Profile resolved =
-        motionTree.hasAnyOverride() ? motionTree.overlayChainOnto(profilePath, base) : base;
-    // Rule override (top of the cascade): a per-window Timing / Curve rule for
-    // this (window, event) replaces the resolved curve / duration. Skipped for
-    // windowless events (desktop switch) and when no rules are configured.
-    if (query.hasWindow() && !m_shaderManager.animationRuleSet().isEmpty()) {
-        resolved = PlasmaZones::resolveAnimationMotionProfile(m_shaderManager.animationRuleEvaluator(), resolved, query,
-                                                              profilePath, windowId, m_curveRegistry);
-    }
-    // Clamp the resolved duration into the animation envelope HERE, at the one
-    // place every consumer shares. The motion tree hands a node's duration
-    // through unclamped (ProfileTree does no bounding, and Profile::fromJson
-    // accepts any finite positive value up to one hour), and the tree is rebuilt
-    // from hand-editable profile JSON.
-    //
-    // The two shader consumers re-clamp the DURATION downstream via
-    // resolveTransitionLifetimeMs. The animator calls that helper too, but only
-    // for the spring maxLifetimeMs cap — its PARAMETRIC duration goes from
-    // applyWindowGeometry straight into WindowAnimator::startAnimation, whose
-    // own clampProfile bounds to [0, 10000] ms, a different, looser envelope.
-    // Without this a `"duration": 5000` node would run a 2 s shader leg on
-    // window.open but a 5 s animator leg on a snap, with its durationMs == 0
-    // shader riding along and pinning per-frame repaints for the full 5 s.
-    // Clamping at the source keeps all three consumers on one envelope; the
-    // downstream clamps then become idempotent.
-    if (resolved.duration) {
-        resolved.duration =
-            static_cast<qreal>(qBound(PhosphorAnimation::Limits::MinAnimationDurationMs, qRound(*resolved.duration),
-                                      PhosphorAnimation::Limits::MaxAnimationDurationMs));
-    }
-    return resolved;
-}
-
-void PlasmaZonesEffect::slotMotionProfileTreeChanged()
-{
-    // A per-event animation duration was edited (daemon rescanned a
-    // `profiles/*.json` override). Re-fetch so per-event durations apply
-    // live, without a logout/login. loadCachedSettings() also re-fetches
-    // it on settingsChanged; this dedicated path covers the profile-file
-    // edits that deliberately do NOT ride settingsChanged.
-    loadMotionProfileTreeFromDbus();
-}
-
-void PlasmaZonesEffect::slotSessionIdleChanged(bool idle)
-{
-    if (m_sessionIdle == idle) {
-        return;
-    }
-    m_sessionIdle = idle;
-    if (!m_pauseAnimationWhenIdle) {
-        // Track the state anyway — the setting can be turned on mid-session, and
-        // the next paint should already know whether we are idle.
-        return;
-    }
-    if (!idle) {
-        // Waking. A paused chain emits no damage of its own, so nothing would put
-        // it back in the paint loop without this.
-        repaintAllDecorations();
-    }
-    // Going idle needs no repaint: the windows simply stop being driven and keep
-    // presenting the composite they already hold.
-}
-
-void PlasmaZonesEffect::loadShaderRegistryFromDbus()
-{
-    // The key is named INLINE, not bound to a local alias first. An alias saves nothing
-    // here and costs real safety: the registry-contract test scrapes these call sites to
-    // prove every key the effect fetches is registered daemon-side, and an alias forces
-    // it to resolve an identifier back to a constant — which it has now got wrong three
-    // separate times, each time silently checking the wrong key while its own self-check
-    // balanced. Name the constant where it is used and there is nothing to resolve.
-    PhosphorProtocol::ClientHelpers::loadSettingAsync(
-        this, PhosphorProtocol::Service::SettingProperty::AnimationShaderSearchPaths, [this](const QVariant& v) {
-            dispatchJsonSetting(PhosphorProtocol::Service::SettingProperty::AnimationShaderSearchPaths, v,
-                                /*objectSink=*/{}, [this](const QJsonArray& arr) {
-                                    QStringList paths;
-                                    for (const auto& entry : arr) {
-                                        if (entry.isString())
-                                            paths.append(entry.toString());
-                                    }
-                                    if (!paths.isEmpty()) {
-                                        m_shaderManager.m_animationShaderRegistry.addSearchPaths(paths);
-                                    }
-                                    qCDebug(lcEffect)
-                                        << "loadShaderRegistryFromDbus: added" << paths.size()
-                                        << "search paths — registry effect count="
-                                        << m_shaderManager.m_animationShaderRegistry.availableEffects().size();
-                                });
-        });
 }
 
 } // namespace PlasmaZones
